added debug switch to build.sh

+++ build.sh
@@ -1,5 +1,4 @@
 #!/bin/bash -e
-
 ARG=""
 if [ ! -z "$TOOLCHAIN_PATH" ]; then
 	ARG="-DTOOLCHAIN_PATH=$TOOLCHAIN_PATH"
@@ -8,7 +7,7 @@ fi
 TYPES="mini big"
 TYPES="big"
 VERSIONS="0_9_5 0_9_6"
-
+ENABLE_DEBUG=0
 NAME=libsensors
 
 rm -rf libs/
@@ -18,10 +17,12 @@ for type in $TYPES; do
 	for ver in $VERSIONS; do
 		mkdir -p libs/RoboCORE_${ver}_${type}
 
-		rm -rf build_tmp/
-		cmake -DHFRAMEWORK_PATH=$HFRAMEWORK_PATH -DROBOCORE_VERSION=$ver -DROBOCORE_TYPE=$type $ARG -H. -Bbuild_tmp
-		make -C build_tmp/
-		cp build_tmp/${NAME}d.a libs/RoboCORE_${ver}_${type}/${NAME}d.a
+		if [ "$ENABLE_DEBUG" = "1" ]; then
+			rm -rf build_tmp/
+			cmake -DHFRAMEWORK_PATH=$HFRAMEWORK_PATH -DROBOCORE_VERSION=$ver -DROBOCORE_TYPE=$type $ARG -H. -Bbuild_tmp
+			make -C build_tmp/
+			cp build_tmp/${NAME}d.a libs/RoboCORE_${ver}_${type}/${NAME}d.a
+		fi
 
 		rm -rf build_tmp/
 		cmake -DHFRAMEWORK_PATH=$HFRAMEWORK_PATH -DROBOCORE_VERSION=$ver -DROBOCORE_TYPE=$type $ARG -DRELEASE=1 -H. -Bbuild_tmp
